requirements.txt
langchain>=0.3
langchain-community>=0.3
langchain-openai>=0.3
langchain-core>=0.3
langchain-text-splitters>=0.3
langgraph>=0.2
openai>=1.0
pypdf>=4.0
streamlit>=1.35
python-dotenv>=1.0
